@@ -134,9 +134,13 @@ func (c *counter) Add(v float64) {
 		return
 	}
 
-	atomicUpdateFloat(&c.valBits, func(oldVal float64) float64 {
-		return oldVal + v
-	})
+	for {
+		oldBits := atomic.LoadUint64(&c.valBits)
+		newBits := math.Float64bits(math.Float64frombits(oldBits) + v)
+		if atomic.CompareAndSwapUint64(&c.valBits, oldBits, newBits) {
+			return
+		}
+	}
 }
 
 func (c *counter) AddWithExemplar(v float64, e Labels) {
