@@ -120,9 +120,13 @@ func (g *gauge) Dec() {
 }
 
 func (g *gauge) Add(val float64) {
-	atomicUpdateFloat(&g.valBits, func(oldVal float64) float64 {
-		return oldVal + val
-	})
+	for {
+		oldBits := atomic.LoadUint64(&g.valBits)
+		newBits := math.Float64bits(math.Float64frombits(oldBits) + val)
+		if atomic.CompareAndSwapUint64(&g.valBits, oldBits, newBits) {
+			return
+		}
+	}
 }
 
 func (g *gauge) Sub(val float64) {
