@@ -1,4 +1,4 @@
-// Copyright 2024 The Prometheus Authors
+// Copyright 2023 The Prometheus Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,29 +11,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build darwin && !cgo
+//go:build wasip1 || js || ios
+// +build wasip1 js ios
 
 package prometheus
 
-func getMemory() (*memoryInfo, error) {
-	return nil, notImplementedErr
+func canCollectProcess() bool {
+	return false
 }
 
-// describe returns all descriptions of the collector for Darwin.
+func (c *processCollector) processCollect(ch chan<- Metric) {
+	c.errorCollectFn(ch)
+}
+
+// describe returns all descriptions of the collector for wasip1 and js.
 // Ensure that this list of descriptors is kept in sync with the metrics collected
 // in the processCollect method. Any changes to the metrics in processCollect
 // (such as adding or removing metrics) should be reflected in this list of descriptors.
 func (c *processCollector) describe(ch chan<- *Desc) {
-	ch <- c.cpuTotal
-	ch <- c.openFDs
-	ch <- c.maxFDs
-	ch <- c.maxVsize
-	ch <- c.startTime
-
-	/* the process could be collected but not implemented yet
-	ch <- c.rss
-	ch <- c.vsize
-	ch <- c.inBytes
-	ch <- c.outBytes
-	*/
+	c.errorDescribeFn(ch)
 }
